+++ packages/expo-image/ios/ImageView.swift
@@ -1,7 +1,7 @@
 // Copyright 2022-present 650 Industries. All rights reserved.
 
-import SDWebImage
 import ExpoModulesCore
+import SDWebImage
 import VisionKit
 
 typealias SDWebImageContext = [SDWebImageContextOption: Any]
@@ -21,8 +21,8 @@ public final class ImageView: ExpoView {
   )
 
   var loadingOptions: SDWebImageOptions = [
-    .retryFailed, // Don't blacklist URLs that failed downloading
-    .handleCookies // Handle cookies stored in the shared `HTTPCookieStore`
+    .retryFailed,  // Don't blacklist URLs that failed downloading
+    .handleCookies,  // Handle cookies stored in the shared `HTTPCookieStore`
   ]
 
   var sources: [ImageSource]?
@@ -43,6 +43,9 @@ public final class ImageView: ExpoView {
 
   var recyclingKey: String? {
     didSet {
+      guard oldValue != nil else {
+        return
+      }
       if recyclingKey != oldValue {
         sdImageView.image = nil
       }
@@ -117,7 +120,8 @@ public final class ImageView: ExpoView {
 
     // Modify URL request to add headers.
     if let headers = source.headers {
-      context[SDWebImageContextOption.downloadRequestModifier] = SDWebImageDownloaderRequestModifier(headers: headers)
+      context[SDWebImageContextOption.downloadRequestModifier] =
+        SDWebImageDownloaderRequestModifier(headers: headers)
     }
 
     context[.cacheKeyFilter] = createCacheKeyFilter(source.cacheKey)
@@ -171,7 +175,7 @@ public final class ImageView: ExpoView {
     onProgress([
       "loaded": isPhotoLibraryAsset ? nil : receivedSize,
       "total": isPhotoLibraryAsset ? nil : expectedSize,
-      "progress": Double(receivedSize) / Double(expectedSize)
+      "progress": Double(receivedSize) / Double(expectedSize),
     ])
   }
 
@@ -199,8 +203,8 @@ public final class ImageView: ExpoView {
           "url": imageUrl?.absoluteString,
           "width": image.size.width,
           "height": image.size.height,
-          "mediaType": imageFormatToMediaType(image.sd_imageFormat)
-        ]
+          "mediaType": imageFormatToMediaType(image.sd_imageFormat),
+        ],
       ])
 
       let scale = window?.screen.scale ?? UIScreen.main.scale
@@ -249,7 +253,8 @@ public final class ImageView: ExpoView {
    Same as `bestSource`, but for placeholders.
    */
   var bestPlaceholder: ImageSource? {
-    return getBestSource(from: placeholderSources, forSize: bounds.size, scale: screenScale) ?? placeholderSources.first
+    return getBestSource(from: placeholderSources, forSize: bounds.size, scale: screenScale)
+      ?? placeholderSources.first
   }
 
   /**
@@ -280,7 +285,8 @@ public final class ImageView: ExpoView {
     // Some loaders (e.g. blurhash) need access to the source.
     context[ImageView.contextSourceKey] = placeholder
 
-    imageManager.loadImage(with: placeholder.uri, context: context, progress: nil) { [weak self] placeholder, _, _, _, finished, _ in
+    imageManager.loadImage(with: placeholder.uri, context: context, progress: nil) {
+      [weak self] placeholder, _, _, _, finished, _ in
       guard let self = self, let placeholder = placeholder, finished else {
         return
       }
@@ -305,7 +311,7 @@ public final class ImageView: ExpoView {
   private func createTransformPipeline() -> SDImagePipelineTransformer {
     let transformers: [SDImageTransformer] = [
       SDImageBlurTransformer(radius: blurRadius),
-      SDImageTintTransformer(color: imageTintColor)
+      SDImageTintTransformer(color: imageTintColor),
     ]
     return SDImagePipelineTransformer(transformers: transformers)
   }
@@ -398,7 +404,9 @@ public final class ImageView: ExpoView {
 
   var enableLiveTextInteraction: Bool = false {
     didSet {
-      guard #available(iOS 16.0, *), oldValue != enableLiveTextInteraction, ImageAnalyzer.isSupported else {
+      guard #available(iOS 16.0, *), oldValue != enableLiveTextInteraction,
+        ImageAnalyzer.isSupported
+      else {
         return
       }
       if enableLiveTextInteraction {
@@ -416,7 +424,9 @@ public final class ImageView: ExpoView {
     }
 
     Task {
-      guard let imageAnalyzer = Self.imageAnalyzer, let imageAnalysisInteraction = findImageAnalysisInteraction() else {
+      guard let imageAnalyzer = Self.imageAnalyzer,
+        let imageAnalysisInteraction = findImageAnalysisInteraction()
+      else {
         return
       }
       let configuration = ImageAnalyzer.Configuration([.text, .machineReadableCode])